Added the option to pass a client/server name to the generation scripts
This does not mean the *.info files need no longer be edited if they also
contain the same name.

The names used in the *.info files and the names passed as an optional
argument to the generation scripts can be totally different if this is
desired.

The *.info files describe information to be included within the generated
files, while the optional argument just modifies the file names to be
generated.